trying to fix the code

+++ imdb.py
@@ -48,7 +48,7 @@ def prepare_data(seqs, labels, MAXLEN_to_pad_to, maxlen=None):
     num_layers_to_pad = MAXLEN_to_pad_to - maxlen
 
     # columns are the samples in R^maxlen
-    x = np.zeros((MAXLEN_to_pad_to, n_samples)).astype('int64')
+    x = np.zeros((MAXLEN_to_pad_to, n_samples)).astype(np.float32)
     x_mask = np.zeros((MAXLEN_to_pad_to, n_samples)).astype(np.float32)
     for idx, s in enumerate(seqs):
         x[:lengths[idx], idx] = s
